refactor(task): add type_traits.hpp
